// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/company-hierarchy-model.ts
export class CompanyHierarchyModel {
    entityId: string;
    entityName: string;
    isEntity: boolean;
    isGroup: boolean;
    isBranch: boolean;
    isCountry: boolean;
    isHeadOffice: boolean;
    timeStamp: boolean;
    parentEntityId: string;
    childEntityId: string;
    currencyId: string;
    timeZoneId: string;
    countryId: string;
    countryName: string;
    isArchive: boolean;
    defaultPayrollTemplateId: string;
    address: string;
    description: string;
    operationsPerformedBy: string;
    city: string;
    street: string;
    state: string;
    postalCode: string;
}